+++ print.h
@@ -1,0 +1,2 @@
+
+void print(lis_obj * obj);
